lambda/requirements.txt
simsimd==3.7.6
boto3 == 1.34.40
langchain == 0.1.6
pinecone-client == 3.0.2
langchain_pinecone == 0.0.1
langchain_openai == 0.0.5

